+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/workflow_with_specified_events_to_dispatch.php
@@ -9,7 +9,6 @@ $container->loadFromExtension('framework', [
         'my_workflow' => [
             'type' => 'state_machine',
             'marking_store' => [
-                'type' => 'method',
                 'property' => 'state',
             ],
             'supports' => [